Feat Tank Bullet System

+++ Assets/Game/Scripts/Interfaces/IDamageable.cs
@@ -1,5 +1,4 @@
-using UnityEngine;
 public interface IDamageable
 {
-    void TakeDamage(float ammount);
+    void TakeDamage(int ammount);
 }
